Complete test for 'flip' function factory
