+++ tests/functional_tests/flip_tests/test_properties.py
@@ -1,0 +1,33 @@
+from typing import (Any,
+                    Callable,
+                    Dict)
+
+from lz.functional import (curry,
+                           flip)
+from lz.hints import Domain
+
+
+def test_involution(transparent_function: Callable[..., Any],
+                    transparent_function_args: Domain,
+                    transparent_function_kwargs: Dict[str, Domain]) -> None:
+    double_flipped = flip(flip(transparent_function))
+
+    original_result = transparent_function(*transparent_function_args,
+                                           **transparent_function_kwargs)
+    double_flipped_result = double_flipped(*transparent_function_args,
+                                           **transparent_function_kwargs)
+
+    assert double_flipped_result == original_result
+
+
+def test_currying(transparent_function: Callable[..., Any],
+                  transparent_function_args: Domain,
+                  transparent_function_kwargs: Dict[str, Domain]) -> None:
+    flipped = flip(transparent_function)
+    curried_flipped = curry(flipped)
+
+    result = curried_flipped(*transparent_function_args,
+                             **transparent_function_kwargs)
+
+    assert result == flipped(*transparent_function_args,
+                             **transparent_function_kwargs)
